Use common fixture
